Add option to save favicon files in a subfolder
I would like to suggest the following changes.
I added a third option to the constructor: $subfolder.

If you supply a subfolder to the constructor, the favicons will be saved in that subfolder of /favicon.

Reason I needed this because I developed one central codebase that is used by a multitude of websites. They need to be able to save a favicon set for each individual website.

Current setup of the package doesn't support this as the favicons will always be saved in the same /favicon folder.